feat: Add Emoji Bitmap over the faces in the pictures
The whichEmoji() method already determines the proper Emoji based on the facial expression.

You should now use it in a switch statement to select the proper emoji drawable. Do the following:

Change the return type of the whichEmoji() method to the Emoji enum.

At the end of the method, return the proper Emoji.

Rename the detectFaces() method to detectFacesAndOverlayEmoji() since this method will now overlay the proper drawable as well as detect the faces, and change it's return type from void to Bitmap. Initialize an empty drawable inside the loop which iterates through the detected faces called emojiBitmap, used to hold the correct emoji drawable. Create a switch statement using the result of the whichEmoji() method as the argument, with a case for each Emoji, and use the BitmapFactory.decodeResource() method to define the initialized drawable to be the appropriate Emoji based on the result of the switch.

You should now have the following components: The original picture bitmap from the camera, passed into the detectFacesAndOverlayEmoji() method. The appropriate Emoji bitmap for each face, defined inside the faces for loop. The detected face, again defined inside the loop.

The next step involves creating a new method, addBitmapToFace(), which takes the three components we just discussed (picture bitmap, Emoji bitmap, and Face object), and combines them, overlaying the Emoji bitmap over the detected face on each iteration of the loop and returning the resulting, combined bitmap.

Combining Bitmaps is straightforward, but outside of the scope of this lesson. You can copy the code for this method below, and I will step through the functionality in the next node. For a more detailed explanation on working with 2D graphics take a look at the Developer Docs.